requirements.txt
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
filelock==3.8.0
huggingface-hub==0.10.0
idna==3.4
numpy==1.23.3
packaging==21.3
PyAudio==0.2.12
pycparser==2.21
pyparsing==3.0.9
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
srt==3.5.2
tokenizers==0.12.1
torch==1.12.1
tqdm==4.64.1
transformers==4.22.2
typing_extensions==4.4.0
urllib3==1.26.12
vosk==0.3.44
websockets==10.3
